frontend-utils: Move Bookmark from desktop
