+++ frontend-utils/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod bookmarks;
 pub mod parse;
